refactor: move `Rhino` implementation to `evaluator` package
