+++ src/test/kotlin/com/rapatao/projects/ruleset/engine/cases/TestData.kt
@@ -1,7 +1,7 @@
 package com.rapatao.projects.ruleset.engine.cases
 
 import com.rapatao.projects.ruleset.engine.context.EvalEngine
-import com.rapatao.projects.ruleset.engine.rhino.RhinoEvalEngine
+import com.rapatao.projects.ruleset.engine.evaluator.rhino.RhinoEvalEngine
 import org.junit.jupiter.params.provider.Arguments
 import java.math.BigDecimal
