+++ src/test/kotlin/com/rapatao/projects/ruleset/engine/evaluator/rhino/ParserTest.kt
@@ -1,4 +1,4 @@
-package com.rapatao.projects.ruleset.engine.rhino
+package com.rapatao.projects.ruleset.engine.evaluator.rhino
 
 import com.rapatao.projects.ruleset.engine.types.Expression
 import com.rapatao.projects.ruleset.engine.types.Operator
